New test StampHeader.testSandeepSinghHeaderTable() which focuses on the stackoverflow question
table header in pdf getting displayed using itextpdf5.1.1 but not in
itextpdf5.5.3
http://stackoverflow.com/questions/29977927/table-header-in-pdf-getting-displayed-using-itextpdf5-1-1-but-not-in-itextpdf5-5

Indeed, the code as presented by the OP does not show the header table.
This makes sense, though:

The OP has cells with default padding (i.e. 2) and height 10, and he
tries to insert text at height 7. But 2 (top margin) + 7 (text height) +
2 (bottom margin) = 11, i.e. more than fits into the cell height 10.
Thus, the text does not fit and is not displayed.

You can fix this by either

* using a smaller font, e.g. 6, or
* using a heigher cell, e.g. 11, or
* using a smaller padding, e.g. 1.