New test ImportPageWithoutFreeSpace focusing on the stackoverflow question
iText: Importing styled Text and informations from an existing PDF
http://stackoverflow.com/questions/31980979/itext-importing-styled-text-and-informations-from-an-existing-pdf

The method testImportPages demonstrates how to import merely the region
of a PDF page with actual content. The main necessity is to call
cropPdf(PdfReader) for the reader in question which restricts the media
boxes of the pages to the bounding box of the existing content.


Fixed NullPointerException situation in
MarginFinder.renderPath(PathPaintingRenderInfo) which became obvious
while developing the test.