New test CreateEmptyField.testDA3161_Template focusing on the stackoverflow question
New Signature Field Rotated 90 Degrees
http://stackoverflow.com/questions/32332490/new-signature-field-rotated-90-degrees
DA3161-Template.pdf
https://drive.google.com/open?id=0B3pKBz-WDrXnM3hYeDFtXzhldnM

The page in your document is rotated using the Rotate page dictionary
entry. When creating a field to be filled-in by others later, you have
to add a hint to the field indicating a counter-rotation if you want the
field content to effectively appear upright.

You do this by setting the MKRotation attribute of the field. This
creates a rotation entry R with value 90 in the appearance
characteristics dictionary MK of the field.