New test RemoveHeaderFooter focusing on the stackoverflow question
Issue in Removing Header and Footer in PDF using iText PDF
http://stackoverflow.com/questions/32448118/issue-in-removing-header-and-footer-in-pdf-using-itext-pdf
spec.pdf
https://www.dropbox.com/s/xznwx4ogemsgd42/spec.pdf?dl=0

testSpec65_65 applies the OP's method cleanUpContent to his test file
removing 65 units wide header and footer section. According to the OP
this results in empty pages and exceptions.

Using iText 5.5.6 there indeed is an exception, and the page during the
cleanup of which the exception occured, is blank. Using the current
5.5.7-SNAPSHOT, though, no exception is thrown and no page is blanked.
Issues can be observed, though, on landscape pages of the test file.

testSpec45_45 applies the OP's method cleanUpContent to his test file
removing 45 units wide header and footer section. According to the OP
this works.

Using iText 5.5.6, though there is an exception, and the page during the
cleanup of which the exception occured, is blank. Using the current
5.5.7-SNAPSHOT, on the other hand, no exception is thrown and no page is
blanked. Issues can be observed, though, on landscape pages of the test
file.