New test ReadForm focusing on the stackoverflow question
iTextsharp : AcroFields - Unable to cast object of type
'iTextSharp.text.pdf.PdfDictionary' to type
'iTextSharp.text.pdf.PdfArray'
http://stackoverflow.com/questions/32455178/itextsharp-acrofields-unable-to-cast-object-of-type-itextsharp-text-pdf-pdf
Sample.PDF
https://drive.google.com/file/d/0B3W8aJry8ZMERnJubHpMdVk5SmM/view?usp=sharing

Indeed, parsing the form definition results in a class cast error. The
cause for that exception is that the AcroForm PDF form description in
your sample PDF is invalid: In the AcroForm interactive form dictionary
the value of the Fields key is a dictionary object but that value must
be an array.