New test ExtractSuperAndSubInLine focussing on the stackoverflow question
How can I extract subscript / superscript properly from a PDF using
iTextSharp?
http://stackoverflow.com/questions/33492792/how-can-i-extract-subscript-superscript-properly-from-a-pdf-using-itextsharp
310cmr7.pdf
http://www.mass.gov/courts/docs/lawlib/300-399cmr/310cmr7.pdf

This test tests the TextLineFinder and HorizontalTextExtractionStrategy
with the beginning of the OP's sample document.

TextLineFinder is a RenderListener which tries to identify horizontal
text lines by projecting the text bounding boxes onto the y axis. It
assumes that these projections do not overlap for text from different
lines, even in case of subscripts and superscripts. This class
essentially is a reduced form of the PageVerticalAnalyzer.

HorizontalTextExtractionStrategy is a TextExtractionStrategy which uses
a TextLineFinder to identify horizontal text lines and then uses these
informations to sort the text chunks.