New test AnalyzeSignatures focusing on the stackoberflow question
PDF signing with timestamp: certificate details does not appear in
timestamp properties
http://stackoverflow.com/questions/34544380/pdf-signing-with-timestamp-certificate-details-does-not-appear-in-timestamp-pro
PDFSigned.pdf, the signature being extracted as
"PDFSigned.pdf.Signature1.raw"
https://drive.google.com/file/d/0B61KdyZ43x-9cEk5YVV6WVd4WUU/view?usp=sharing

The OP used the same certificate for signing the signature and the time
stamp, it in particular does not have the required extended key usage
marking it as a time stamping certificate.

The test uses the new helper class SignatureAnalyzer. This class is
meant to eventually become a tool for analyzing signatures. More and
more tests shall be added to indicate the issues of the upcoming test
signatures.