New test CreateLink focusing on the stackoverflow question
Create local link in rotated PdfPCell in iTextSharp
http://stackoverflow.com/questions/34408764/create-local-link-in-rotated-pdfpcell-in-itextsharp

This is the equivalent Java code for the C# code in the question.
Indeed, this code also gives rise to the broken result. The cause is
simple: Normally iText does not touch the current transformation matrix.
So the chunk link creation code assumes the current user coordinate
system to be the same as used for positioning annotations. But in case
of rotated cells iText does change the transformation matrix and
consequently the chunk link creation code positions the annotation at
the wrong location.