New test RemoveSignature focusing on the iText-questions question
trying to remove a signature from pdf file
http://itext.2136553.n4.nabble.com/trying-to-remove-a-signature-from-pdf-file-tt4660983.html
PDFSignedFirmaCerta.pdf
http://itext.2136553.n4.nabble.com/attachment/4660983/0/PDFSignedFirmaCerta.pdf

Indeed, this code fails with a NullPointerException. The cause is that a
dubious construct created by the signature software then is processed by
iText code not sufficiently defensively programmed: The signature claims
to have an annotation on a page but that page does claim not to have any
anotations at all.