New TextExtraction test testLocation_text_extraction_test focusing on the stackoverflow question
iText: Extracted text from pdf file using LocationTextExtractionStrategy
is in wrong order
http://stackoverflow.com/questions/35344982/itext-extracted-text-from-pdf-file-using-locationtextextractionstrategy-is-in-w
location_text_extraction_test.pdf
https://www.dropbox.com/s/kl2s6038u51gx2q/location_text_extraction_test.pdf?dl=0

Indeed, the LocationTextExtractionStrategy returns the headers in the
wrong order. This is due to slightly different y coordinates of them.

The HorizontalTextExtractionStrategy2 returns the headers and actually
the whole table correctly. Unfortunately it fails where there are
overlapping lines in side-by-side columns, in this case e.g. for the
invoice recipient address.