New test DoubleSpace focusing on the stackoverflow question
Double space not being preserved in PDF
http://stackoverflow.com/questions/35699167/double-space-not-being-preserved-in-pdf

Indeed, the double space collapses into a single one when
copying&pasting from the generated PDF displayed in Adobe Reader. On the
other hand the gap for the double space is twice as wide as for the
single space. So this essentially is a quirk of copy&paste of Adobe
Reader (and some other PDF viewers, too).