New test EditPageContent focusing on the stackoverflow question
Removing Watermark from PDF iTextSharp
http://stackoverflow.com/questions/35526822/removing-watermark-from-pdf-itextsharp
test3.pdf
https://www.dropbox.com/s/qvlo1v9uzgpu4nj/test3.pdf?dl=0

This class tests the PdfContentStreamEditor and
TransparentGraphicsRemover classes.

PdfContentStreamEditor is the base editor classes which by default acts
as the identity operation (or at least it produces an equivalent content
stream).

TransparentGraphicsRemover on the other hand changes all vector graphics
drawing operations so that operations which would create somewhat
transparent result are dropped. This suffices to not draw the watermark
in the OP's sample PDF.

Beware: This latter sample editor is very simple:

* It only considers transparency created by the ExtGState parameters ca
and CA, it in particular ignores masks.
* It does not look for operations saving or restoring the graphics
state.

These limitations can easily be lifted but require more code than
appropriate for a stackoverflow answer.