New TextExtraction test testTest_pdf focusing on the stackoverflow question
Text from PDF parsing differently using iText
http://stackoverflow.com/questions/37307289/text-from-pdf-parsing-differently-using-itext
test pdf.pdf
https://github.com/Baddy247/Shared-Files/raw/master/test%20pdf.pdf

While the rows look like they are they are at a constant level each,
they actually are not. The 'XXX...:' and 'TOTAL :' parts are at y
coordinates 469.45, 457.95, and 446.45 while the '#..', '1', and '2'
parts are at y coordinates 468.65, 457.15, and 445.65.

To consider horizontal text to be on the same line, iText text
extraction using the default text extraction strategy
(LocationTextExtractionStrategy) requires the y coordinates to be the
same after casting to int. (Actually this is somewhat simplified, for
the whole picture look at
LocationTextExtractionStrategy.TextChunkLocationDefaultImp).

In the case at hand this only is the case for the middle row.

The HorizontalTextExtractionStrategy2 on the other hand recognized each
row as a single line.