New TextExtraction test testMonitor_2016_99 focusing on the stackoverflow question
itextpdf insert space beetwen 7 and dot after extract text
http://stackoverflow.com/questions/37439613/itextpdf-insert-space-beetwen-7-and-dot-after-extract-text
monitor_2016_99.pdf
http://185.49.12.119/~pogdan/7spacedot/monitor_2016_99.pdf

iText inserts spaces whenever there is a gap between two consecutive
text chunks which is larger than a certain amount, or if two consecutive
text chunks overlap. It does so to signal that the chunks do not follow
each other in a normal way.

In case of this document a dot following a seven often is moved left as
far as possible without touching the seven so that the character
bounding boxes overlap.