New RedactText test testRedact20150325101924_2102000595_SessionReport focusing on the stackoverflow question
Attempt to apply redactions results in exception
http://stackoverflow.com/questions/37713112/attempt-to-apply-redactions-results-in-exception
20150325101924-2102000595-SessionReport.pdf
https://drive.google.com/file/d/0Bz0Wye-k7GsZUXdKbWtMTUNpcXc/view?usp=sharing

This corresponds to the iTextSharp/C# code by the OP having fixed a
first issue, the use of annots added in the same PdfStamper sesion.

In Java no more exception occurs but in C# one does. This is due to
iText using a HashMap and iTextSharp using a Dictionary as a member of
the PdfCleanUpProcessor to which annotation rectangles are added by
their annotations index in their respective page annotations array
(which is a logical error on both sides). HashMaps allow overwriting
entries, Dictionaries don't. Thus, iTextSharp hickups and iText does
not.