New test DrawGradient focusing on the stackoverflow question
iText gradient issue in landscape
http://stackoverflow.com/questions/39072316/itext-gradient-issue-in-landscape

The problem is that while itext content adding functionalities do take
the page rotation into account (they translate the given coordinates so
that in the rotated page x goes right and y goes up and the origin is in
the lower left), the shading pattern definitions (which are not part of
the page content but externally defined) don't.

Thus, you have to make the shading definition rotation aware, e.g. like
in this drawSexyTriangle implementation.