New test SplitSmart focusing on the stackoverflow question
Remain only used font subsets while splitting PDF in Java
http://stackoverflow.com/questions/39348420/remain-only-used-font-subsets-while-splitting-pdf-in-java
cjk-with-fonts.pdf
https://www.dropbox.com/s/mqgozcbs7k4m40z/cjk-with-fonts.pdf?dl=0

Indeed, splitting results in files the summed-up size of which is
considerably larger than the original size. The reason are no
unnecessary unused entries in the resource dictionaries, there simply
are merely a handful of fonts in the file and each page uses 3 or 4 of
them. In the split PDFs, therefore, most fonts are copied to multiple
files which makes the sum of the sizes of the split files so much larger
than the size of the original file.

One would have to implement a preprocessor which would walk the pages,
xobjects, etc., determine for each of them the actually used glyphs from
each font, create subset fonts and replace the currently associated
fonts by those subset fonts. This is not easy, most probably beyond the
scope of a stackoverflow answer, but should be possible using the
building bricks iText offers.