New TextExtraction tests testPb and testPbNoToUnicode focusing on a comment to the stackoverflow question
Extract text with iText not works: encoding or crypted text?
http://stackoverflow.com/questions/37748346/extract-text-with-itext-not-works-encoding-or-crypted-text
pb.pdf
https://dl.dropboxusercontent.com/u/6413030/pb.pdf

The document has not been provided by the OP but by sschuberth
(http://stackoverflow.com/users/1127485/sschuberth) in a comment.

Indeed, testPb shows that this text is not properly extracted.

In contrast, though, testPbNoToUnicode first removes the ToUnicode
tables of the fonts. And indeed, now extraction succeeds.