New ChangeTextColor test testChangeBlackTextToGreenDocument focusing on the stackoverflow question
Traverse whole PDF and change some attribute with some object in it using iText
http://stackoverflow.com/questions/40401800/traverse-whole-pdf-and-change-some-attribute-with-some-object-in-it-using-itext

This test shows how to change the color of text of a given color. In this case, black text is changed to green.

Beware, this is a proof-of-concept, not a final and complete solution. In particular

* Text is considered to be black if for its color the expression BaseColor.BLACK.equals(color) is true; as equality among BaseColor and its descendant classes is not completely well-defined, this might lead to some false positives.
* PdfContentStreamEditor only inspects and edits the content stream of the page itself, not the content streams of displayed form xobjects or patterns; thus, some text may not be found.

Improving the class to properly detect black color and to recursively traverse and edit the content streams of used patterns and xobjects remains as an exercise for the reader.