New SetCheckBox tests focusing on the stack overflow question
itext fill checkbox with value
http://stackoverflow.com/questions/42819618/itext-fill-checkbox-with-value
PD3100.pdf
https://www.poreskaupravars.org/Documents/Doc/PD3100.pdf

testPd3100SetVrPr4ToNo uses a valid name and so at least creates a PDF displayed with a somehow selected field. 

testPd3100SetVrPr4ToNoAppend uses a valid name in append mode and so creates a PDF displayed in Adobe Reader with a selected field as desired.