New CreateEllipse tests focusing on the stack overflow question
Trying to draw an ellipse annotation and the border on the edges goes thin and thik when i try to roatate pdf itext5
http://stackoverflow.com/questions/43205385/trying-to-draw-an-ellipse-annotation-and-the-border-on-the-edges-goes-thin-and-t

testCreateEllipse creates an ellipse annotation without appearance on a page without rotation. Everything looks ok.

testCreateEllipseAppearance creates an ellipse annotation with appearance on a page without rotation. Everything looks ok.

testCreateEllipseOnRotated creates an ellipse annotation without appearance on a page with rotation. The ellipse form looks ok but it is moved to the right of the actual appearance rectangle when viewed in Adobe Reader. This is caused by iText creating a non-standard rectangle, the lower left not being the lower left etc.

testCreateEllipseAppearanceOnRotated creates an ellipse annotation with appearance on a page with rotation. The ellipse position looks ok but it is deformed. This is caused by iText rotating the annotation rectangle but not (how could it?) the appearance rectangle.

testCreateCorrectEllipseAppearanceOnRotated creates an ellipse annotation with appearance with switched dimensions on a page with rotation. Everything looks ok.