New ExtractHash test focusing on the stack overflow question
PDF Signing, generated PDF Document certification is invalid? (using external signing, web-eid, HSM)
https://stackoverflow.com/questions/44196316/pdf-signing-generated-pdf-document-certification-is-invalid-using-external-si
signed_file.pdf
https://www.mediafire.com/?fqvnf9mg50pfzjh

The new test method testPareshSignedFile shows that the hash in the DigestInfo (cf. the answer to the question above) is not identical to the actual hash of the signed ranges.

The helper method extractHashes has even been extended to also calculate the hash of the document hash. This also is not the signed hash in the case at hand.