New CreateSignature test focusing on the stack overflow question
Text - PDFAppearence issue
https://stackoverflow.com/questions/45062602/itext-pdfappearence-issue

This test shows how one can create a custom signature layer 2.

As the OP of the question at hand mainly wants to generate a pure DESCRIPTION appearance that uses the whole area, we here essentially copy the PdfSignatureAppearance.getAppearance code for generating layer 2 in pure DESCRIPTION mode and apply it to a plain pre-fetched layer 2.