New classes and tests focusing on the stack overflow question
How to find all occurrences of specific text in a PDF and insert a page
break above?
https://stackoverflow.com/questions/45823741/how-to-find-all-occurrences-of-specific-text-in-a-pdf-and-insert-a-page-break-ab

SearchTextLocationExtractionStrategy is a text extraction strategy that
finds all occurrences of specific text in a PDF, it actually even allows
the use of a regular expression.
TextLocationExtraction tests show how to find all occurrences of
specific text in a PDF using the SearchTextLocationExtractionStrategy
and mark them.

AbstractPdfPageSplittingTool is an abstract utility class that inserts a
page break above a custom set of vertical positions.
SplitPages test testSplitDocumentInThirds shows how to split the pages
of a document in three equal parts using the
AbstractPdfPageSplittingTool.

SplitPages test testSplitDocumentAboveAngestellter shows how to find all
occurrences of specific text in a PDF using the
SearchTextLocationExtractionStrategy and then insert a page break above
using the AbstractPdfPageSplittingTool.