New Merging test focusing on the stack overflow question
when i open a pdf in adobe acrobat pro Dc the text is getting messed up
for pdf creation i used itext5.5.12
https://stackoverflow.com/questions/46120706/when-i-open-a-pdf-in-adobe-acrobat-pro-dc-the-text-is-getting-messed-up-for-pdf

Test files "1 Loyds.pdf", "2 CRPWLI.pdf", "3 SLC Dec.pdf", "4
Schedule.pdf", and "5 Sched of INS.pdf" were received via e-mail from
Kishore Rachakonda (kishore.rachakonda@yash.com) on 2017-09-11 17:31.

This test is a port of the Ruby-on-Rails merge routine provided by the
OP. The result does not have the problem described by the OP.

Later it became clear that the OP's merge result was post-processed by
at least two other programs, and one of those post processors appears to
have optimized the use of embedded fonts. Unfortunately "5 Sched of
INS.pdf" is not a completely valid PDF: It uses an embedded subset of a
font but does not mark the font name accordingly. Thus, the optimizing
post processor added this mere font subset (assuming it to be the whole
font program) to font resources which require glyphs missing in the
subset.