New VerifySignature test testVerifyTestDsp focusing on the stack overflow question
Digital Signature Verification with itext not working
https://stackoverflow.com/questions/46346144/digital-signature-verification-with-itext-not-working
test_dsp.pdf
https://drive.google.com/open?id=0B1XKjvoeoyPZWnk5bzc5T3VSQUk

The issue is that the signature uses ECDSA and iText 5 does not (yet) support ECDSA. "Support" here actually means that iText cannot find the name ECDSA for the OID 1.2.840.10045.4.3.2 (SHA256withECDSA) to build a proper algorithm name to use for verification.

Adding a mapping "1.2.840.10045.4.3.2" to "ECDSA" resolves the issue.