New TextExtraction test focusing on the stack overflow question
Extract text from pdf using itextsharp returns empty string
https://stackoverflow.com/questions/47869635/extract-text-from-pdf-using-itextsharp-returns-empty-string
Smlouva ZŠ Komenského 66 NJ - pronájem učebny.pdf
https://smlouvy.gov.cz/smlouva/soubor/5376672/Smlouva%20Z%C5%A0%20Komensk%C3%A9ho%2066%20NJ%20-%20pron%C3%A1jem%20u%C4%8Debny.pdf

The PDF itself is weird to start with: It appears to have been OCR'ed ignoring the page rotation, i.e. effectively with the content rotated by 180°, so the recognized text is weird.

But as testSmlouvaZSKomenskeho66NJpronajemucebny shows there also is an issue of iText: It does not properly extract the so recognized text.

This probably is due to the font of the text not having a ToUnicode map and furthermore having a mixed single-byte/double-byte encoding.