New VerifyTimestamp test focusing on the stack overflow question
iText pdf timestamp validation returns false, why?
https://stackoverflow.com/questions/48211757/itext-pdf-timestamp-validation-returns-false-why
testpdf_timestamp.pdf
https://drive.google.com/file/d/1skI3NM9cqw2m6eW9jKXaJXKzvCjyQMib/view

The code the OP used for inspiration was for retrieving information from a signature which may include a signature time stamp. The PDF of the OP, on the other hand, contains a document time stamp. The call pkcs7.verifyTimestampImprint() checks the time stamp as a signature time stamp and, therefore, fails.