New AnalyzeSignatures test focusing on the stack overflow question
C# PKCS7 Smartcard Digital Signature - Document has been altered or corrupted since it was signed
https://stackoverflow.com/questions/51031446/c-sharp-pkcs7-smartcard-digital-signature-document-has-been-altered-or-corrupt
signedpdf.pdf, the signature being extracted as "signedpdf.pdf.dsa.raw".
https://drive.google.com/file/d/1I6hUO8B3fnkgws9Pk1Puti9HjkRJE0Rq/view?usp=sharing

As it turns out, the signature DigestInfo object contains the hash of the hash of the signed attributes, not simply the hash of the signed attributes, i.e. the signed attributes incorrectly are hashed twice.

The SignatureAnalyzer class has been extended to catch this.