New CreateDocumentLikeJavaaac test focusing on the stack overflow question
Pdfwriter: document has no content
https://stackoverflow.com/questions/52131139/pdfwriter-document-has-no-content

This tests the inner class Pdfgenerator which essentially is the OP's code with constants inserted for unknown expression values.

As it turns out, the code in its main path automatically must throw an IndexOutOfBoundsException, see the first lines of the loop in createTable, and the other path adds nearly no content at all to the document.