New TestRotatedImagePosition tests focusing on the stack overflow question
iText image.setRotationDegrees() not keeping a consistent origin
https://stackoverflow.com/questions/52153094/itext-image-setrotationdegrees-not-keeping-a-consistent-origin

testPositionsLikeSujayJ is the original code of the OP. Indeed, the rotation center to get one image from the other is not an obvious corner.

The reason: The Image is not positioned at the given coordinates and then rotated around an obvious point. Instead the Image is rotated, then its bounding box is determined, and finally this bounding box is positioned to have its lower left at the given coordinates.

This is more obvious with more rectangles, cf. testPositionsLikeSujayJAndMore.