New AddImageOverTable test focusing on the stack overflow question
Add image above table - Itext
https://stackoverflow.com/questions/53095527/add-image-above-table-itext

testLikeWotonSampaio tests the OP's code.
Indeed, the table lines always were on top of the image, no matter which
was added first. The cause was that bitmap images via
Document#add(Element) are added to a virtual layer under text and table
data.
The solution was to add the image to PdfWriter.getDirectContent()
instead of to the Document, cf. TemplatePDF.addImg(int, float, float,
float).