New CreateLink test focusing on the stack overflow question
Unicode symbols in iTextSharp relative link
https://stackoverflow.com/questions/53264518/unicode-symbols-in-itextsharp-relative-link

This test checks how to create a link to a local file with non-ASCII
characters in its path. Unfortunately iText 5.x ignores the possibility
that special characters might be contained in the anchor argument,
applies the standard PDF document encryption, and actually drops all
characters not in that encoding.

The two options working properly are either injecting the UTF-8 encoded
URI or URL encoding the URI beforehand. Actually the former method makes
use of a PDF-2 specific feature, UTF-8 encoded strings, so the latter
one may be preferable.

Probably the Launch Action variant can be made working, too, by finding
the proper encoding to inject the file path with.