New CreateSignatureComodo tests focusing on the stack overflow question
How to enable LTV for a timestamp signature and set the pdf change not
allowed?
https://stackoverflow.com/questions/53297489/how-to-enable-ltv-for-a-timestamp-signature-and-set-the-pdf-change-not-allowed

testCreateSimpleSignature creates a simple signature.

testCreateNoChangesAllowedCertificationAndLtv creates a
no-changes-allowed certification and thereafter adds LTV information.
Adobe does not accept this in spite of using a PDF-2.0 file or a PDF-1.7
file with applicable developer extensions.

testCreateCertificationAndLtvAndNoChangesAllowed creates a form-filling
certification, thereafter adds LTV information, and finally signs again
making use of a signature lock dictionary that causes a FieldMDP
transform that reduces the access to no-changes-allowed. Adobe accepts
this but the incremental update can easily be removed and the remains
would be form-filling allowed again.

testCreateLtvNoChangesAllowedCertification creates a no-changes-allowed
certification and in that process also adds LTV information. Adobe does
accept this.