New AnalyzeSignatures test focusing on the stack overflow question
How to compute PDF signature hash?
https://stackoverflow.com/questions/54361974/how-to-compute-pdf-signature-hash
dummy-signed.pdf, the signature being extracted as
"dummy-signed.pdf.Signature2.raw"
https://gist.githubusercontent.com/nowox/94dd54e484df877e1232c18bd7b91c97/raw/d249f3757137e9b665e895c900f08b1156f1bc4f/dummy-signed.pdf.base64

As it turns out, the OP looked at the wrong hash value (not the hash
value in the messageDigest signed attribute but the encrypted hash value
in the signature bytes). The value in the messageDigest signed attribute
coincides with the hash of the signed byte ranges of the example PDF.