New AnalyzeSignatures test focusing on the esig DSS issue
DSS-1538 - Cannot analyze signatures Exception Error
https://ec.europa.eu/cefdigital/tracker/browse/DSS-1538
vypis_z_kn.pdf, the signature being extracted as
"vypis_z_kn.pdf.Signature1.raw"
https://ec.europa.eu/cefdigital/tracker/secure/attachment/18729/vypis_z_kn.pdf

There is no issue validating the signature. As it turns out the problem
was that the PDF is encrypted and the signature value dictionary does
not have a type entry. Thus, PDFBox does not recognize the unencrypted
signature container as such and attempts to decrypt it, scrambling it in
the course of that.