New PhrasesChunksAndFonts test focusing on the stack overflow question
How can I use regular and bold in a single String using iText Android
https://stackoverflow.com/questions/54784126/how-can-i-use-regular-and-bold-in-a-single-string-using-itext-android

If you set a font in the Paragraph constructor, the font used for a
later added Chunk object is the font of that chunk supplemented by data
from the paragraph font in properties not set in the chunk font.

The style of the font is a bit field and unfortunately this
supplementing in the style field is implemented by means of bit-wise
or-ing. Thus, the BOLD flag from the paragraph font is or-ed to the
styles of all the chunks added to the paragraph!