New AddNotFittingImage tests focusing on the stack overflow question
How to add the words in a new page while adding images and words in itext
https://stackoverflow.com/questions/55630476/how-to-add-the-words-in-a-new-page-while-adding-images-and-words-in-itext

The tests in this test class analyze how iText behaves when an image is added to a document which does not fit on the current page, and it examines different strategies to tweak this default behavior, using newPage() calls and/or the property StrictImageSequence of PdfWriter.

iText default behavior: Image is floated to the next page but paragraph drawn after image remains on current page. (testAddNotFittingImage)

iText default behavior with a newPage() call after adding the image: Image is floated to the next page and paragraph drawn after that call comes immediately after the image. (testAddNotFittingImageNewPage)

iText default behavior with 2 newPage() calls after adding the image: Image is floated to the next page and paragraph drawn after that call comes after the image on yet another new page. (testAddNotFittingImageDoubleNewPage)

iText behavior with StrictImageSequence: A new page is generated on which the image is drawn, the paragraph drawn after image follows the image immediately if possible. (testAddNotFittingImageStrictSequence)

iText behavior with StrictImageSequence with a newPage() call after adding the image: A new page is generated on which the image is drawn, the paragraph drawn after image follows the image on yet another new page. (testAddNotFittingImageStrictSequenceNewPage)