New RedactWithImageIssue tests focusing on the stack overflow question
What makes PdfStamper to remove images from pdf after cleanup() though it shouldn't?
https://stackoverflow.com/questions/56186498/what-makes-pdfstamper-to-remove-images-from-pdf-after-cleanup-though-it-should
bad.pdf, good.pdf
https://drive.google.com/drive/folders/1buvmQCiHhrbVD_i0mJhrdCo3YM9ngL3p?usp=sharing

testRedactBadLikeShatterwoods - Indeed, when redacting this "bad" PDF the image vanishes without it being in the redaction areas.
testRedactGoodLikeShatterwoods - Indeed, when redacting this "good" PDF the image does not vanishes.

testRedactPdfWithInlineImages - As it turned out, in the "bad" test PDF of the OP the unexpectedly vanishing image is an inline image completely outside all redaction areas. This test illustrates the issue even more clearly using a PDF containing only 5 inline images. iText redaction applied with the readaction area covering the third image completely, the second and fourth partially, and the first and fifth not at all, results in the first and fifth image vanishing.