New CloudBoxAnnotation tests focusing on the stack overflow question
Box Cloud Annotation Appearance in iText
https://stackoverflow.com/questions/56388137/box-cloud-annotation-appearance-in-itext

testDrawLikeChitgoks - Indeed, this code creates a broken visualization. But the created content stream is already invalid, with arbitrary instructions between path definition and path drawing, and furthermore there is the instruction `app.rectangle(cborder.getBBox())` essentially doing nothing which surely is not meant like this.

testDrawLikeChitgoksImproved - This code fixes the issue in the OP's original code: It makes the appearance stream valid by removing the arbitrary instructions between path definition and path drawing, and it properly applies the bounding box of the cloud to the appearance.