New CreateMeasureDictionary tests focusing on wen li's answer to the stack overflow question
Create measure dictionary for area annotation in iText
https://stackoverflow.com/questions/56667651/create-measure-dictionary-for-area-annotation-in-itext

testCreateLikeChitgoksArea - Indeed, with this code one gets a polygon Adobe Reader measures as 0.04 cm². As it turns out, the conversion factor calculation by the OP was incorrect.

testCreateLikeChitgoksAreaImproved - There were two errors in the conversion factor calculated in the OP's original code: The factor for the areas was used as is as conversion factor along each axis, and the original polygon area was calculated incorrectly. This has been corrected here.