New TimingXmlWorker tests focusing on the stack overflow question
Is there an efficient way to write several HTML strings to a PDF document in Java?
https://stackoverflow.com/questions/57043863/is-there-an-efficient-way-to-write-several-html-strings-to-a-pdf-document-in-jav

testMakePdfLikeEvanV - The time this takes is clearly lower than the time observed by the OP. The difference might be explainable by very bad equipment.

testMakePdfLikeEvanVSingleWorkerCall - In contrast to testMakePdfLikeEvanV in this test the HTML pieces are glued together and then transformed using only a single XMLWorker call. I had to drastically increase the number of "fields" to get a runtime above one second! Thus the observation by the OP that this change doesn't improve the runtime cannot be explained by bad equipment anymore, there must be an unknown factor acting as a break.