New RemovePageTextContent tests focusing on the stack overflow question
Replace text inside a PDF file using iText
https://stackoverflow.com/questions/57308588/replace-text-inside-a-pdf-file-using-itext

PdfContentStreamEditor - The base class has been enhanced to offer a finalizeContent method called when the end of the content is reached. Furthermore, it has a new constructor allowing to set a custom RenderListener instead of the default dummy.

SimpleTextRemover - This content stream editor illustrates what is necessary to properly remove text from a PDF. Actually it is limited in a few aspects...

testRemoveTestFromTest - This test applies the SimpleTextRemover to a simple test document and removes its occurrences of "Test".

testRemovestteFromTest - This test applies the SimpleTextRemover to a simple test document and removes its occurrences of "stte". The document text ends with "test", i.e. with a partial match. Thus, this test checks whether final ignoring of a partial match and flushing of the remaining cached operations works properly. At the same time, though, it illustrates that the remover ignores gaps because it recognizes a match in the middle of "test test".

testRemoveéFromTestSteveB - This test applies the SimpleTextRemover to a test document and removes its occurrences of "é". The document makes use of text showing operators with side effects (namely '). Thus, this test checks whether the remover properly adds the former side effects when replacing ' by TJ.