New CreateFooter tests focusing on the stack overflow question
How to add page number on the footer of pdf using itext which should take care of its width?
https://stackoverflow.com/questions/57577981/how-to-add-page-number-on-the-footer-of-pdf-using-itext-which-should-take-care-o

testCreateFooterLikeShashiShekhar - This test essentially contains the OP's original code. Indeed, already for two-digit page numbers the left cell, "Page x of", breaks into two lines. The cause is that the left column is too small to host more than single-digit page numbers.

testCreateFooterLikeShashiShekharImproved - This test improves the OP's original code a bit by widening the first column of the footer table and right-aligning it to prevent an unnatural gap between "Page x of" and the total number of pages.