New CreateStamp test focusing on the OpenPDF issue 253
Issue with createStamp annotation on Adobe Acrobat Reader DC
https://github.com/LibrePDF/OpenPDF/issues/253

testCreateMultipleLikeDprutean - In contrast to what the OP claims to
have found, i.e. that the size of the template changes the position of
the yellow balloon icon, the output of this test shows no influence of
the template size on that position.