New MakeLtvEnabled test focusing on the stack overflow question
Add Timestamp for hardware token digital signature and adding LTV not
working / throws exception
https://stackoverflow.com/questions/59469148/add-timestamp-for-hardware-token-digital-signature-and-adding-ltv-not-working

testSampleECDSA - The direct reason for the exception is that iText's
PdfPKCS7 class does not support ECDSA. This test confirmed this.

As a consequence the AdobeLtvEnabling class has been changed to not rely
on PDFPKCS7 but instead use BC code directly.

Beware: Due to an issue of the OCSP certificates of the PKI in question this test now recurses forever. This still has to be fixed.