New CreateSignature tests focusing on the stack overflow question
Multiple Signings in pdf File using IText
https://stackoverflow.com/questions/62271473/multiple-signings-in-pdf-file-using-itext

signLikePauloGonçalvesOriginal - This is like the OP's original code. Already the first sign call truncates the original file before iText could read it. Thus, iText throws an exception.

signLikePauloGonçalvesEdited - This is like the OP's edited code. Both sign calls sign the original, unsigned document. Thus, the remaining output is the file with the signature from the second signing call.

signLikePauloGonçalvesCorrected - This code uses appropriate file streams for the both signing calls. In particular the output of the first signing is the input of the second one, and in no case a signing call uses the same file for input and output.
  Furthermore, the improved sign method used here allows the use of distinct signature field names for the calls and refrains from setting inappropriate certification levels.