New AddImageInBetween test focusing on the stack overflow question 66352174
Background image itextpdf 5.5
https://stackoverflow.com/questions/66352174/background-image-itextpdf-5-5

UnderContentRemover - This content stream editor removes the former UnderContent from the edited content stream and provides it as list of operations. UnderContent is recognized as a leading block of instructions enveloped in a save-graphics-state/restore-graphics-state frame.

testForVladimirSafonov - This test applies the UnderContentRemover to test data generated using the code provided by the OP and uses it to add a stamp between the former UnderContent and DirectContent.