New OpenFile tests focusing on the stack overflow question 66885882
Writing a streamed cross-reference in PDF: file is detected as damaged?
https://stackoverflow.com/questions/66885882/writing-a-streamed-cross-reference-in-pdf-file-is-detected-as-damaged
corrupted-pdf.pdf
https://easyupload.io/4z0gfc

testCorruptedPdf - In this version the first issue is the incorrect startxref offset. For pure cross reference stream PDFs iText cannot repair the cross references as it expects to find a trailer. But see testCorruptedPdfStartxref.

testCorruptedPdfStartxref - After fixing the incorrect startxref offset, see testCorruptedPdf(), the actual problem pops up: the predictor bytes are missing in the stream data. For pure cross reference stream PDFs iText cannot repair the cross references as it expects to find a trailer.